functions/serverless-security/requirements.txt
google-cloud-storage
google-cloud-kms